test: change to follow GitLab's testing best practices
https://docs.gitlab.com/ee/development/testing_guide/best_practices.html
