+++ spec/cli_spec.rb
@@ -21,9 +21,8 @@ RSpec.describe Miteru::CLI do
     it "should download file(s)" do
       url = "http://#{host}:#{port}/has_kit"
       zip_files = ["test.zip"]
-
       expect(Dir.glob("#{base_dir}/*.zip").empty?).to be(true)
-      capture(:stdout) { subject.download_zip_files(url, zip_files, @path) }
+      capture(:stdout) { subject.download_zip_files(url, zip_files, base_dir) }
       expect(Dir.glob("#{base_dir}/*.zip").empty?).to be(false)
     end
   end
